Refactor dashboard into submodules
